requirements.txt
asgiref==3.8.1
boto3==1.39.9
botocore==1.39.9
certifi==2025.7.14
charset-normalizer==3.4.2
diff-match-patch==20241021
dj-database-url==3.0.1
Django==5.2.1
django-anymail==13.0
django-cors-headers==4.7.0
django-import-export==4.3.8
django-jazzmin==3.0.1
django-storages==1.14.6
django-tinymce==4.1.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.10
environs==14.2.0
gunicorn==24.1.1
idna==3.10
inflection==0.5.1
jmespath==1.0.1
marshmallow==4.0.0
packaging==25.0
pillow==11.3.0
psycopg2-binary==2.9.11
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2023.3.post1
PyYAML==6.0.2
requests==2.32.4
s3transfer==0.13.1
shortuuid==1.0.13
six==1.17.0
sqlparse==0.4.4
stripe==12.3.0
tablib==3.8.0
typing_extensions==4.9.0
tzdata==2025.3
uritemplate==4.1.1
urllib3==1.26.18
whitenoise==6.11.0
